moved Firebase.js file
